requirements.txt
boto3
langchain 
streamlit 
pillow 
faiss-cpu